+++ src/main/java/com/lsoria/qrsec/domain/dto/SimplifiedGuestDTO.java
@@ -1,0 +1,14 @@
+package com.lsoria.qrsec.domain.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+public class SimplifiedGuestDTO {
+
+    private String dni;
+
+}